requirements.txt
vpyTelegramBotAPI==4.15.0
Pillow==10.0.0
requests==2.31.0
google-generativeai==0.3.0